Update the profile
