+++ package.json
@@ -20,7 +20,7 @@
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --detectOpenHandles ",
+    "test": "react-scripts test--inspect-brk",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
